django/pyproject.toml
[tool.poetry]
name = "django-godot-websockets"
version = "0.1.0"
description = ""
authors = ["Forest Anderson <forestkzanderson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
channels = "^3.0.3"
channels-redis = "^3.2.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
